Create using phyton
This code uses Selenium to:

Open a new Chrome browser instance
Navigate to YouTube
Wait for the search box to be clickable
Enter a search query
Click the search button
Wait for the search results to load
Print the titles of the search results
Close the browser
Note that you'll need to replace "path/to/chromedriver" with the actual path to the chromedriver executable on your system. You can download chromedriver from the official Chromium website.

Also, make sure to replace "your search query here" with the actual search query you want to enter.Here's a breakdown of the code:

The first three lines import the necessary modules from the Selenium library:
webdriver is used to create a new browser instance
By is used to specify the strategy for finding elements on the page (e.g. by name, by ID, etc.)
WebDriverWait and EC (Expected Conditions) are used to wait for specific conditions to be met on the page (e.g. an element to be clickable)
The next line creates a new Chrome browser instance using the webdriver module
The driver.get() method navigates to the specified URL (in this case, YouTube)
The WebDriverWait and EC modules are used to wait for specific elements to be present on the page (e.g. the search box, the search button, etc.)
The send_keys() method is used to enter text into the search box
The click() method is used to click the search button
The find_element_by_name() method is used to find elements on the page by their name attribute
The print() function is used to print the titles of the search results
Finally, the driver.quit() method is used to close the browser instance
Python is a popular language for web scraping and automation tasks because of its ease of use, flexibility, and extensive libraries (like Selenium).